# T-398 Source Map Performance enhancements