--- a/packages/core/core/package.json
+++ b/packages/core/core/package.json
@@ -23,7 +23,7 @@
     "@parcel/logger": "^2.0.0-alpha.3.1",
     "@parcel/package-manager": "^2.0.0-alpha.3.1",
     "@parcel/plugin": "^2.0.0-alpha.3.1",
-    "@parcel/source-map": "^2.0.0-alpha.4",
+    "@parcel/source-map": "^2.0.0-alpha.4.3",
     "@parcel/types": "^2.0.0-alpha.3.1",
     "@parcel/utils": "^2.0.0-alpha.3.1",
     "@parcel/workers": "^2.0.0-alpha.3.1",