--- a/packages/core/core/src/InternalAsset.js
+++ b/packages/core/core/src/InternalAsset.js
@@ -89,7 +89,7 @@ type InternalAssetOptions = {|
   value: Asset,
   options: ParcelOptions,
   content?: Blob,
-  map?: ?SourceMap,
+  mapBuffer?: ?Buffer,
   ast?: ?AST,
   idBase?: ?string,
 |};
@@ -98,7 +98,6 @@ export default class InternalAsset {
   value: Asset;
   options: ParcelOptions;
   content: Blob;
-  map: ?SourceMap;
   mapBuffer: ?Buffer;
   ast: ?AST;
   idBase: ?string;
@@ -107,14 +106,14 @@ export default class InternalAsset {
     value,
     options,
     content,
-    map,
+    mapBuffer,
     ast,
     idBase,
   }: InternalAssetOptions) {
     this.value = value;
     this.options = options;
     this.content = content || '';
-    this.map = map;
+    this.mapBuffer = mapBuffer;
     this.ast = ast;
     this.idBase = idBase;
   }
@@ -154,12 +153,12 @@ export default class InternalAsset {
           }),
         ),
       ),
-      this.map == null
+      this.mapBuffer == null
         ? Promise.resolve()
         : this.options.cache.setBlob(
             this.getCacheKey('map' + pipelineKey),
-            // $FlowFixMe
-            this.mapBuffer ? this.mapBuffer : this.map.toBuffer(),
+            // $FlowFixMe strange that this occurs
+            this.mapBuffer,
           ),
     ]);
     this.value.contentKey = contentKey;
@@ -216,29 +215,30 @@ export default class InternalAsset {
   }
 
   async getMapBuffer(): Promise<?Buffer> {
-    if (!this.mapBuffer) {
-      if (this.map) {
-        this.mapBuffer = this.map.toBuffer();
-      } else if (this.value.mapKey != null) {
-        this.mapBuffer = await this.options.cache.getBlob(this.value.mapKey);
-      }
+    if (!this.mapBuffer && this.value.mapKey != null) {
+      this.mapBuffer = await this.options.cache.getBlob(this.value.mapKey);
     }
 
     return this.mapBuffer;
   }
 
   async getMap(): Promise<?SourceMap> {
-    if (!this.map && this.value.mapKey != null) {
-      // Get sourcemap from flatbuffer
-      this.map = new SourceMap();
-      this.map.addBufferMappings(await this.getMapBuffer());
+    if (!this.mapBuffer) {
+      await this.getMapBuffer();
     }
 
-    return this.map;
+    if (this.mapBuffer) {
+      // Get sourcemap from flatbuffer
+      let map = new SourceMap();
+      map.addBufferMappings(this.mapBuffer);
+      return map;
+    }
   }
 
   setMap(map: ?SourceMap): void {
-    this.map = map;
+    if (map) {
+      this.mapBuffer = map.toBuffer();
+    }
   }
 
   getCacheKey(key: string): string {
@@ -328,7 +328,7 @@ export default class InternalAsset {
       options: this.options,
       content,
       ast: result.ast,
-      map: result.map,
+      mapBuffer: result.map ? result.map.toBuffer() : null,
       idBase: this.idBase,
     });