--- a/packages/core/core/src/public/Asset.js
+++ b/packages/core/core/src/public/Asset.js
@@ -156,6 +156,10 @@ class BaseAsset {
   getMap(): Promise<?SourceMap> {
     return this.#asset.getMap();
   }
+
+  getMapBuffer(): Promise<?Buffer> {
+    return this.#asset.getMapBuffer();
+  }
 }
 
 export class Asset extends BaseAsset implements IAsset {