--- a/packages/core/source-map/src/SourceMap.js
+++ b/packages/core/source-map/src/SourceMap.js
@@ -1,2 +1,0 @@
-// @flow
-export {default, generateEmptyMap} from 'source-maps';