--- a/packages/core/types/index.js
+++ b/packages/core/types/index.js
@@ -316,6 +316,7 @@ export interface BaseAsset {
   getBuffer(): Promise<Buffer>;
   getStream(): Readable;
   getMap(): Promise<?SourceMap>;
+  getMapBuffer(): Promise<?Buffer>;
   getIncludedFiles(): $ReadOnlyArray<File>;
   getDependencies(): $ReadOnlyArray<Dependency>;
   getConfig(