--- a/packages/optimizers/terser/src/TerserOptimizer.js
+++ b/packages/optimizers/terser/src/TerserOptimizer.js
@@ -41,14 +41,15 @@ export default new Optimizer({
     };
 
     let sourceMap = null;
+    let mappingsBuffer = [];
     if (options.sourceMaps) {
       sourceMap = new SourceMap();
+
       // $FlowFixMe
       config.output = {
         source_map: {
           add(source, gen_line, gen_col, orig_line, orig_col, name) {
-            // $FlowFixMe
-            sourceMap.addMapping({
+            mappingsBuffer.push({
               source,
               name,
               original: {
@@ -60,16 +61,23 @@ export default new Optimizer({
                 column: gen_col,
               },
             });
+
+            if (mappingsBuffer.length > 25) {
+              // $FlowFixMe
+              sourceMap.addIndexedMappings(mappingsBuffer);
+              mappingsBuffer = [];
+            }
           },
         },
       };
     }
 
-    if (sourceMap && map) {
-      sourceMap = await map.extend(sourceMap);
-    }
-
     let result = minify(contents, config);
+
+    if (sourceMap && map) {
+      sourceMap.addIndexedMappings(mappingsBuffer);
+      sourceMap = sourceMap.extends(map.toBuffer());
+    }
 
     if (result.error) {
       throw result.error;