--- a/packages/optimizers/terser/src/TerserOptimizer.js
+++ b/packages/optimizers/terser/src/TerserOptimizer.js
@@ -35,6 +35,7 @@ export default new Optimizer({
       },
       sourceMap: {
         filename: path.relative(options.projectRoot, bundle.filePath),
+        asObject: true,
       },
       module: bundle.env.outputFormat === 'esmodule',
     };