--- a/packages/transformers/react-refresh-wrap/src/ReactRefreshWrapTransformer.js
+++ b/packages/transformers/react-refresh-wrap/src/ReactRefreshWrapTransformer.js
@@ -118,9 +118,7 @@ export default new Transformer({
 
       res.code = generated.code;
       // $FlowFixMe...
-      res.map = new SourceMap(generated.rawMappings, {
-        [sourceFileName]: null,
-      });
+      res.map = new SourceMap(generated.rawMappings);
     }
 
     if (asset.meta.globals && asset.meta.globals.size > 0) {