--- a/packages/transformers/react-refresh-wrap/src/ReactRefreshWrapTransformer.js
+++ b/packages/transformers/react-refresh-wrap/src/ReactRefreshWrapTransformer.js
@@ -117,8 +117,13 @@ export default new Transformer({
       );
 
       res.code = generated.code;
+      let map = null;
+      if (generated.rawMappings) {
+        map = new SourceMap();
+        map.addIndexedMappings(generated.rawMappings);
+      }
       // $FlowFixMe...
-      res.map = new SourceMap(generated.rawMappings);
+      res.map = map;
     }
 
     if (asset.meta.globals && asset.meta.globals.size > 0) {