… packages/core/source-map/test/add-map.test.js
… packages/core/source-map/test/extend-sourcemap.test.js
… packages/core/source-map/test/get-positions-sourcemap.test.js
… packages/examples/kitchen-sink/package.json
… packages/optimizers/terser/package.json
… packages/packagers/js/package.json
… packages/transformers/babel/package.json
… packages/transformers/coffeescript/package.json
… packages/transformers/js/package.json
… packages/transformers/react-refresh-wrap/package.json
… packages/transformers/typescript-types/package.json
… yarn.lock