<!---
Thanks for filing a pull request 😄 ! Before you submit, please read the following:

Search open/closed issues before submitting since someone might have pushed the same thing before!
-->

# ↪️ Pull Request

<!---
Provide a general summary of the pull request here
Please look for any issues that this PR resolves and tag them in the PR.
-->

This PR switches to the C++ package for sourcemap processing.

## Benchmark

The benchmark shows that the big difference in performance is that serialisation does not need to happen at all as it's passing along buffers which are super-fast compared to json.

| Description | Before | After |
| --- | --- | --- |
| Serialization | 3secs | 175ms |
| addMappings | 2secs | 1,4secs |

## Future Improvements

From my benchmarks and tests I think we could potentially use the vlq mappings instead of flatbuffers for caching, they appear to be only slightly slower and would perhaps allow us to do less processing on sourcemaps in transformers which would probably cut the `addMappings` time in half. But in turn increase serialization times slightly by maybe 50ms or something.

Closes T-398