@mischnic 2020-03-06
What does this mean for the REPL? Can we keep the old sourcemap package and align it's api with the C++ one?