@DeMoorJasper 2020-03-06
@mischnic probably, the api isn't that much different, I tried to keep it as similar as possible. (Probably just alias it in for the repl?)

Did some minor changes however to make it easier and more efficient.