@mischnic 2020-03-06
I get a segfault on Ubuntu, title: "node crashed with SIGSEGV in napi_delete_reference()"