@mischnic 2020-03-06
```
Stacktrace:
 #0  0x0000000000000000 in ?? ()
 No symbol table info available.
 #1  0x00000000009aad8c in napi_delete_reference ()
 No symbol table info available.
 #2  0x00007f059dde2c29 in SourceMapBinding::Init(Napi::Env, Napi::Object) () from /home/niklas/Desktop/parcel-compiled/node_modules/source-maps/prebuilds/linux-x64/node.napi.glibc.node
 No symbol table info available.
 #3  0x00007f059dde346f in __napi_Init(napi_env__*, napi_value__*) () from /home/niklas/Desktop/parcel-compiled/node_modules/source-maps/prebuilds/linux-x64/node.napi.glibc.node
 No symbol table info available.
 #4  0x00000000009c5c04 in napi_module_register_by_symbol(v8::Local<v8::Object>, v8::Local<v8::Value>, v8::Local<v8::Context>, napi_value__* (*)(napi_env__*, napi_value__*)) ()
 No symbol table info available.
 #5  0x00000000009c76de in node::binding::DLOpen(v8::FunctionCallbackInfo<v8::Value> const&)::{lambda(node::binding::DLib*)#1}::operator()(node::binding::DLib*) const ()
 No symbol table info available.
 #6  0x00000000009c6ab2 in node::binding::DLOpen(v8::FunctionCallbackInfo<v8::Value> const&) ()
 No symbol table info available.
 #7  0x0000000000bc9a29 in v8::internal::MaybeHandle<v8::internal::Object> v8::internal::(anonymous namespace)::HandleApiCallHelper<false>(v8::internal::Isolate*, v8::internal::Handle<v8::internal::HeapObject>, v8::internal::Handle<v8::internal::HeapObject>, v8::internal::Handle<v8::internal::FunctionTemplateInfo>, v8::internal::Handle<v8::internal::Object>, v8::internal::BuiltinArguments) ()
 No symbol table info available.
 #8  0x0000000000bcb817 in v8::internal::Builtin_HandleApiCall(int, unsigned long*, v8::internal::Isolate*) ()
 No symbol table info available.
 #9  0x00000000013a72b9 in Builtins_CEntry_Return1_DontSaveFPRegs_ArgvOnStack_BuiltinExit () at ../../deps/v8/../../deps/v8/src/builtins/base.tq:3418
 No locals.
 #10 0x000000000132cb64 in Builtins_InterpreterEntryTrampoline () at ../../deps/v8/../../deps/v8/src/builtins/base.tq:357
 No locals.
```