@DeMoorJasper 2020-03-06
@mischnic awesome, thanks for looking into it. Strange that it doesn't show up when running tests and benchmarks in source-maps repo