@DeMoorJasper 2020-03-07
Strange enough this only happens when using threads, maybe a bug in N-Api? 😞