@mischnic 2020-03-07
I guess that it's rather `node-addon-api`
```
Thread 8 Crashed:
0   node           	0x0000000100049567 napi_delete_reference + 40
1   sourcemap.node 	0x00000001045143ff Napi::Reference<Napi::Function>::Reset() + 47 (napi-inl.h:2271)
2   sourcemap.node 	0x0000000104514380 Napi::Reference<Napi::Function>::operator=(Napi::Reference<Napi::Function>&&) + 32 (napi-inl.h:2191)
3   sourcemap.node 	0x000000010450ebef Napi::FunctionReference::operator=(Napi::FunctionReference&&) + 47 (napi-inl.h:2473)
4   sourcemap.node 	0x000000010450e67b SourceMapBinding::Init(Napi::Env, Napi::Object) + 1483 (SourceMap.cpp:645)
5   sourcemap.node 	0x000000010450ed4d Init(Napi::Env, Napi::Object) + 61 (SourceMap.cpp:653)
6   sourcemap.node 	0x00000001045144ee Napi::RegisterModule(napi_env__*, napi_value__*, Napi::Object (*)(Napi::Env, Napi::Object))::'lambda'()::operator()() const + 94 (napi-inl.h:256)
7   sourcemap.node 	0x000000010451448d napi_value__* Napi::details::WrapCallback<Napi::RegisterModule(napi_env__*, napi_value__*, Napi::Object (*)(Napi::Env, Napi::Object))::'lambda'()>(Napi::RegisterModule(napi_env__*, napi_value__*, Napi::Object (*)(Napi::Env, Napi::Object))::'lambda'()) + 13 (napi-inl.h:69)
8   sourcemap.node 	0x000000010450edee Napi::RegisterModule(napi_env__*, napi_value__*, Napi::Object (*)(Napi::Env, Napi::Object)) + 78 (napi-inl.h:255)
9   sourcemap.node 	0x000000010450ed94 __napi_Init(napi_env__*, napi_value__*) + 36 (SourceMap.cpp:656)
```