@DeMoorJasper 2020-03-07
@mischnic it's the same thing? NAPI = node-addon-api

I found the issue although it's quite complex to fix as far as I can tell. It has to do with worker_threads.  https://nodejs.org/docs/latest-v12.x/api/addons.html#addons_worker_support