@mischnic 2020-03-07
> #### `node-addon-api`
>
> Module for using N-API from C++

The official NAPI itself is C only