@DeMoorJasper 2020-03-07
@mischnic I would also assume there would be some level of warning or error but apparently not, threw a debugger on it and it really points me towards n-api code. I don't actually keep any state it's all n-api unfortunately

![Screenshot 2020-03-07 at 16 07 14](https://user-images.githubusercontent.com/2175521/76145838-c03b7a00-608d-11ea-8b9f-f4a60b8bf8c8.png)