@mischnic 2020-03-07
This seems to work:
```c++
Napi::Object SourceMapBinding::Init(Napi::Env env, Napi::Object exports) {
    Napi::HandleScope scope(env);

    Napi::Function func = DefineClass(env, "SourceMap", {
            InstanceMethod("addRawMappings", &SourceMapBinding::addRawMappings),
            InstanceMethod("addBufferMappings", &SourceMapBinding::addBufferMappings),
            InstanceMethod("stringify", &SourceMapBinding::stringify),
            InstanceMethod("toBuffer", &SourceMapBinding::toBuffer),
            InstanceMethod("getMap", &SourceMapBinding::getMap),
            InstanceMethod("addIndexedMappings", &SourceMapBinding::addIndexedMappings),
            InstanceMethod("addNames", &SourceMapBinding::addNames),
            InstanceMethod("addSources", &SourceMapBinding::addSources),
            InstanceMethod("getSourceIndex", &SourceMapBinding::getSourceIndex),
            InstanceMethod("getNameIndex", &SourceMapBinding::getNameIndex),
            InstanceMethod("extends", &SourceMapBinding::extends),
            InstanceMethod("generateEmptyMap", &SourceMapBinding::generateEmptyMap),
            InstanceMethod("findClosestMapping", &SourceMapBinding::findClosestMapping),
    });

    exports.Set("SourceMap", func);
    return exports;
}

NAPI_MODULE_INIT() {
  return SourceMapBinding::Init(env, Napi::Object(env, exports));
}
```