@DeMoorJasper 2020-03-07
@mischnic was just reading through an issue in node-addon-api and found similar code, does seem to work. Not entirely sure why I had a persistent reference originally but it seems to still work without.

Will do some testing and hopefully get it stable asap