@devongovett 2020-03-22
Looks like flow is failing on missing definitions. Maybe we should add a flow declaration to the source map repo?